+++ custom_components/nikobus/__init__.py
@@ -5,21 +5,27 @@ import logging
 from typing import Any, Final
 
 import voluptuous as vol
-
-from homeassistant.config_entries import ConfigEntryState
-from homeassistant.core import HomeAssistant, ServiceCall, ServiceResponse, SupportsResponse
-from homeassistant.helpers import config_validation as cv, device_registry as dr, entity_registry as er
-from homeassistant.helpers.typing import ConfigType
-from homeassistant.exceptions import ConfigEntryNotReady, ServiceValidationError
 from homeassistant.components import (
-    switch,
-    light,
-    cover,
     binary_sensor,
     button,
+    cover,
+    light,
     scene,
     sensor,
+    switch,
 )
+from homeassistant.config_entries import ConfigEntryState
+from homeassistant.core import (
+    HomeAssistant,
+    ServiceCall,
+    ServiceResponse,
+    SupportsResponse,
+)
+from homeassistant.exceptions import ConfigEntryNotReady, ServiceValidationError
+from homeassistant.helpers import config_validation as cv
+from homeassistant.helpers import device_registry as dr
+from homeassistant.helpers import entity_registry as er
+from homeassistant.helpers.typing import ConfigType
 
 from .const import CONFIG_ENTRY_VERSION, DOMAIN, HUB_IDENTIFIER
 from .coordinator import NikobusConfigEntry, NikobusDataCoordinator